# Conflicts:
#	ParkInspect/ParkInspect/Model/LocalModel/Absence.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Asignment.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Client.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Contactperson.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Datatype.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Employee.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Employee_Status.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Field.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Form.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Formfield.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Image.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Inspection.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Parkinglot.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Region.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Report.cs
#	ParkInspect/ParkInspect/Model/LocalModel/ReportFieldType.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Role.cs
#	ParkInspect/ParkInspect/Model/LocalModel/State.cs
#	ParkInspect/ParkInspect/Model/LocalModel/Template.cs

+++ ParkInspect/ParkInspect/ViewModel/PrepareViewModel.cs
@@ -272,19 +272,28 @@ namespace ParkInspect.ViewModel
         }
         private void SaveDirections()
         {
+            vm.CleanFiles();
             if (directionItems.Count > 0)
             {
                 String runpath = System.IO.Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
                 {
-                    System.IO.StreamWriter SaveFile = new System.IO.StreamWriter(runpath + "/directions/" + _directions_save_name + ".txt");
-                    SaveFile.WriteLine("ID:" + selectedInspection.id);
-                    SaveFile.WriteLine("HOME:" + _home_adress);
-                    foreach (String s in directionItems)
+                    if (!File.Exists((runpath + "/directions/" + _directions_save_name + ".txt")))
                     {
-                        SaveFile.WriteLine(s);
+                        System.IO.StreamWriter SaveFile = new System.IO.StreamWriter(runpath + "/directions/" + _directions_save_name + ".txt");
+
+                        SaveFile.WriteLine("ID:" + selectedInspection.id);
+                        SaveFile.WriteLine("HOME:" + _home_adress);
+                        foreach (String s in directionItems)
+                        {
+                            SaveFile.WriteLine(s);
+                        }
+                        SaveFile.Close();
+                        _dialog.ShowMessage("Succes!", "De routebeschrijving is succesvol opgeslagen!");
                     }
-                    SaveFile.Close();
-                    _dialog.ShowMessage("Succes!", "De routebeschrijving is succesvol opgelsagen!");
+                    else
+                    {
+                        _dialog.ShowMessage("Er ging iets mis!", "Deze naam bestaat al!");
+                    }
                 }
 
                 vm.LoadDirections();
